fix(server): 修改 schema 相关 api，移除 config 模块
