+++ server/src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod collaboration;
 pub mod plugin;
 pub mod program;
+pub mod schema;
 pub mod user;
