--- server/src/config/mod.rs
@@ -1,1 +1,0 @@
-
